Changes needed to get project to compile after admin to server changes
Fixes #31
