+++ pom.xml
@@ -2,7 +2,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>org.springframework.cloud</groupId>
-	<artifactId>spring-cloud-dataflow-admin-mesos-parent</artifactId>
+	<artifactId>spring-cloud-dataflow-server-mesos-parent</artifactId>
 	<version>1.0.0.BUILD-SNAPSHOT</version>
 	<packaging>pom</packaging>
 
@@ -14,8 +14,10 @@
 	</parent>
 
 	<modules>
-		<module>spring-cloud-dataflow-admin-mesos</module>
 		<module>spring-cloud-dataflow-admin-mesos-docs</module>
+		<module>spring-cloud-dataflow-deployer-mesos</module>
+		<module>spring-cloud-dataflow-server-mesos</module>
+		<module>spring-cloud-starter-dataflow-server-mesos</module>
 	</modules>
 
 </project>
