--- spring-cloud-dataflow-admin-mesos/pom.xml
@@ -1,82 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<artifactId>spring-cloud-dataflow-admin-mesos</artifactId>
-	<version>1.0.0.BUILD-SNAPSHOT</version>
-	<packaging>jar</packaging>
-	<name>spring-cloud-dataflow-admin-mesos</name>
-	<description>Data Flow Admin for Mesos/Marathon</description>
-	<parent>
-		<groupId>org.springframework.cloud</groupId>
-		<artifactId>spring-cloud-dataflow-parent</artifactId>
-		<version>1.0.0.BUILD-SNAPSHOT</version>
-		<relativePath />  <!-- lookup parent from repository -->
-	</parent>
-	<properties>
-		<start-class>org.springframework.cloud.dataflow.admin.AdminApplication</start-class>
-		<!-- The following marathon connector version property is inserted in admin.yml -->
-		<spring.cloud.marathon.connector.version>1.0.0.BUILD-SNAPSHOT</spring.cloud.marathon.connector.version>
-		<marathon-connector.client.version>1.0.0.BUILD-SNAPSHOT</marathon-connector.client.version>
-		<spring-cloud-dataflow.version>1.0.0.BUILD-SNAPSHOT</spring-cloud-dataflow.version>
-		<marathon-client.version>0.4.11</marathon-client.version>
-	</properties>
-	<dependencies>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-marathon-connector</artifactId>
-			<version>${marathon-connector.client.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-dataflow-server-core</artifactId>
-			<version>${spring-cloud-dataflow.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>com.cloudbees.marathon</groupId>
-			<artifactId>marathon-client</artifactId>
-			<version>${marathon-client.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-core</artifactId>
-		</dependency>
-		<!--Force Feign versions -->
-		<dependency>
-			<groupId>com.netflix.feign</groupId>
-			<artifactId>feign-core</artifactId>
-			<version>6.1.3</version>
-		</dependency>
-		<dependency>
-			<groupId>com.netflix.feign</groupId>
-			<artifactId>feign-gson</artifactId>
-			<version>6.1.3</version>
-		</dependency>
-		<dependency>
-			<groupId>com.google.code.gson</groupId>
-			<artifactId>gson</artifactId>
-			<version>2.3</version>
-		</dependency>
-	</dependencies>
-	<build>
-		<resources>
-			<resource>
-				<directory>src/main/resources</directory>
-				<filtering>true</filtering>
-				<includes>
-					<include>admin.yml</include>
-					<include>META-INF/spring.factories</include>
-				</includes>
-			</resource>
-		</resources>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-resources-plugin</artifactId>
-			</plugin>
-			<plugin>
-				<groupId>org.springframework.boot</groupId>
-				<artifactId>spring-boot-maven-plugin</artifactId>
-			</plugin>
-		</plugins>
-	</build>
-</project>
