--- spring-cloud-dataflow-admin-mesos/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.springframework.cloud.dataflow.admin.spi.marathon.MarathonAutoConfiguration
