+++ spring-cloud-dataflow-deployer-mesos/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.springframework.cloud.dataflow.server.spi.marathon.MarathonAutoConfiguration
