+++ spring-cloud-dataflow-server-mesos/src/main/resources/dataflow-server.yml
@@ -1,0 +1,36 @@
+server:
+  port: 9393
+management:
+  contextPath: /management
+info:
+  app:
+    name: "@project.artifactId@"
+    description: "@project.description@"
+    version: "@project.version@"
+security:
+  basic:
+    enabled: false
+
+spring:
+  application:
+    name: spring-cloud-dataflow-server-mesos
+  cloud:
+    config:
+      uri: http://localhost:8888
+
+marathon:
+  includes: org.springframework.cloud:spring-cloud-marathon-connector:@spring.cloud.marathon.connector.version@
+
+# If you prefer to use Eureka to locate the Config Server, you can do that by setting
+# spring.cloud.config.discovery.enabled=true (default "false"). The net result of that is 
+# that client apps all need a bootstrap.yml (or an environment variable) with the Eureka 
+# server address, e.g. in eureka.client.serviceUrl.defaultZone
+#
+#  cloud:
+#     config:
+#        discovery:
+#           enabled: true
+#eureka:
+#  client:
+#    serviceUrl:
+#      defaultZone: http://localhost:8761/eureka/
